% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi.R
\name{plot.pi}
\alias{plot.pi}
\title{This funxtion draws a plot representing the approximtion of the number pi above}
\usage{
\method{plot}{pi}(x)
}
\description{
This funxtion draws a plot representing the approximtion of the number pi above
}
