% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDemo.R
\name{runDemo}
\alias{runDemo}
\title{Shiny App for estimating Pi}
\usage{
runDemo()
}
\description{
Shiny App used to estimate pi and plot the corresponding graph
with the different functions estimate_pi and estimate_pi2 and varying the
number of generated points and the seed.
}
\author{
Bron Luca

Germano David

Grandadam Patrik

Lomazzi Vincent

Raisin Edgar
}
